+++ pkg/commands/workload_get.go
@@ -25,11 +25,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	apierrs "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/cli-runtime/pkg/resource"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	cartov1alpha1 "github.com/vmware-tanzu/apps-cli-plugin/pkg/apis/cartographer/v1alpha1"
@@ -39,7 +35,7 @@ import (
 	"github.com/vmware-tanzu/apps-cli-plugin/pkg/completion"
 	"github.com/vmware-tanzu/apps-cli-plugin/pkg/flags"
 	"github.com/vmware-tanzu/apps-cli-plugin/pkg/printer"
-	metav1beta1 "k8s.io/apimachinery/pkg/apis/meta/v1beta1"
+	"github.com/vmware-tanzu/apps-cli-plugin/pkg/source"
 )
 
 type WorkloadGetOptions struct {
@@ -54,8 +50,6 @@ var (
 	_ validation.Validatable = (*WorkloadGetOptions)(nil)
 	_ cli.Executable         = (*WorkloadGetOptions)(nil)
 )
-
-const contentType = "application/json"
 
 func (opts *WorkloadGetOptions) Validate(ctx context.Context) validation.FieldErrors {
 	errs := validation.FieldErrors{}
@@ -226,7 +220,7 @@ func (opts *WorkloadGetOptions) Exec(ctx context.Context, c *cli.Config) error {
 	}
 
 	arg := []string{"Pod"}
-	if tableResult, err := FetchResourceObject(c, workload, arg); err != nil {
+	if tableResult, err := source.FetchResourceObject(c, workload, arg); err != nil {
 		c.Eprintf("\n")
 		c.Eerrorf("Failed to list pods:\n")
 		c.Eprintf("  %s\n", err)
@@ -306,54 +300,4 @@ func areAllResourcesReady(resourcesConditions ...*metav1.Condition) bool {
 		}
 	}
 	return true
-}
-func decodeIntoTable(info []*resource.Info) (runtime.Object, error) {
-	if len(info) == 1 {
-		obj := info[0].Object
-		event, isEvent := obj.(*metav1.WatchEvent)
-		if isEvent {
-			obj = event.Object.Object
-		}
-		if !recognizedTableVersions[obj.GetObjectKind().GroupVersionKind()] {
-			return nil, fmt.Errorf("attempt to decode non-Table object")
-		}
-
-		unstr, ok := obj.(*unstructured.Unstructured)
-		if !ok {
-			return nil, fmt.Errorf("attempt to decode non-Unstructured object")
-		}
-		table := &metav1.Table{}
-		if err := runtime.DefaultUnstructuredConverter.FromUnstructured(unstr.Object, table); err != nil {
-			return nil, err
-		}
-		if len(table.Rows) == 0 {
-			return nil, nil
-		}
-		for i := range table.Rows {
-			row := &table.Rows[i]
-			if row.Object.Raw == nil || row.Object.Object != nil {
-				continue
-			}
-			var converted runtime.Object
-			var err error
-
-			converted, err = runtime.Decode(unstructured.UnstructuredJSONScheme, row.Object.Raw)
-			if err != nil {
-				return nil, err
-			}
-
-			row.Object.Object = converted
-		}
-
-		return table, nil
-	}
-	return nil, nil
-}
-
-// TablePrinter decodes table objects into typed objects before delegating to another printer.
-// Non-table types are simply passed through
-
-var recognizedTableVersions = map[schema.GroupVersionKind]bool{
-	metav1beta1.SchemeGroupVersion.WithKind("Table"): true,
-	metav1.SchemeGroupVersion.WithKind("Table"):      true,
 }
